Remove 'sign' command
