+++ Sources/PackageRegistryTool/PackageRegistryTool.swift
@@ -33,7 +33,6 @@ public struct SwiftPackageRegistryTool: ParsableCommand {
             Login.self,
             Logout.self,
             Publish.self,
-            ArchiveSign.self,
         ],
         helpNames: [.short, .long, .customLong("help", withSingleDash: true)]
     )
